pmci: mctp: samples: Remove unnecessary include
These samples do not seem to need the unistd.h header, but including it
limits us to C libraries which have it, which as it is a POSIX
extension is not all.
So let's not include it so we do not limit ourselves unnecessarily.

(cherry picked from commit 312b1ef04f577d71e6169a47fb5a7ab66cf1cdae)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 312b1ef04f577d71e6169a47fb5a7ab66cf1cdae
Cr-Build-Id: 8695782486137842193
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8695782486137842193
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Id1aee7ebd2d3182a714dffc892311db9d90bd843
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7251289
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>