kernel: poll: patch recursive lock in z_vrfy_k_poll
In z_vrfy_k_poll, there is a memory access check
K_SYSCALL_MEMORY_WRITE which is wrapped in a spinlock, the same
spinlock used in z_handle_obj_poll_events which is called from
k_sem_give() for example.

The K_SYSCALL_MEMORY_WRITE() macro conditionally calls LOG_ERR()
which may call the UART console, which may call an API like
k_sem_give(). This will cause a deadlock since the locked spinlock
will be relocked, and a recursive lock if SPINLOCK_VALIDATE and
ASSERTS are enabled as the validation will fail, causing a LOG_ERR,
causing a k_sem_give() causing a relock... until stack overflows.

To solve the issue, only protect the copy of events to events_copy
with the spinlock, the content of events is not actually checked, and
bound is not shared, so there is no need to do this validation in a
critical section. The contents of events is shared so that must be
copied in atomically.

(cherry picked from commit 6e4ef4484733d19b490743b13243ace697bdb57c)

Original-Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>
GitOrigin-RevId: 6e4ef4484733d19b490743b13243ace697bdb57c
Cr-Build-Id: 8693200467051364689
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8693200467051364689
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3b019ff1380a1a6cf37e26eb44fdb2a517ff3017
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7418627
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>