kernel: O(1) search for threads among CPUs
Instead of performing a linear search to determine if a given
thread is running on another CPU, or if it is marked as being
preempted by a metaIRQ on any CPU do this in O(1) time.

On SMP systems, Zephyr already tracks the CPU on which a thread
executes (or lasted executed). This information is leveraged to
do the search in O(1) time.

(cherry picked from commit 669a8d0704e889302cff10beb0eb419a28282a53)

Original-Signed-off-by: Peter Mitsis <peter.mitsis@intel.com>
GitOrigin-RevId: 669a8d0704e889302cff10beb0eb419a28282a53
Cr-Build-Id: 8693200467051364689
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8693200467051364689
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I785a4239eb778479940da737a813801658290a69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7418629
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>