Bluetooth: Classic: L2CAP: Fix sending buffer reference counting
There is a corner case that the ACL_Data_Packet_Length is less than
the MPS of the L2CAP BR channel connection. Then the partial data
will be sent and the remaining data will be sent in sequence in the
function `l2cap_br_data_pull()`. The issue occurs when sending the
continuing fragment HCI ACL packet, due to the reference counting of
sending buffer is 0. Therefore, the application will be asserted in
the function `send_buf()`.

Fix the issue by managing the reference count of the sending buffer.

Use the function `net_buf_ref()` to increment the reference count of
the sending buffer.

If the sending buffer is the last fragment, use `net_buf_unref()` to
decrement the reference count of the sending buffer to ensure that
the send buffer can be properly released when the data is sent out.

(cherry picked from commit bd7da95049c19405458d437b6c143ab44d6aec33)

Original-Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>
GitOrigin-RevId: bd7da95049c19405458d437b6c143ab44d6aec33
Cr-Build-Id: 8692883383065769921
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8692883383065769921
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ica7b33e760b63e176c1bd9aeff2b6079914a6a46
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7452261
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>